[combobox-announcer] Move types around for package
